renamed jest files
