+++ GeocodingLambda/index.ts
@@ -1,14 +1,12 @@
 /**
  * @type {import('@types/aws-lambda').APIGatewayProxyHandler}
  */
-import "TiFShared"
-
 import {
   conn,
   exponentialFunctionBackoff
 } from "TiFBackendUtils"
-import { LocationCoordinate2D } from "../../TiFShared/domain-models/LocationCoordinate2D.js"
-import { Result, promiseResult, success } from "../../TiFShared/lib/Result.js"
+import { LocationCoordinate2D } from "TiFShared/domain-models/LocationCoordinate2D.js"
+import { Result, promiseResult, success } from "TiFShared/lib/Result.js"
 import { SearchClosestAddressToCoordinates, addLocationToDB, checkExistingPlacemarkInDB, getTimeZone } from "./utils.js"
 
 interface LocationSearchRequest extends LocationCoordinate2D {}
