+++ GeocodingLambda/jest.config.ts
@@ -5,10 +5,10 @@ export default {
     "^.+\\.(t|j)s?$": "@swc/jest"
   },
   transformIgnorePatterns: [
-    "/node_modules/(?!(node-fetch|data-uri-to-buffer|fetch-blob|formdata-polyfill))"
+    "/node_modules/(?!(TiFShared|node-fetch|data-uri-to-buffer|fetch-blob|formdata-polyfill))"
   ],
   moduleNameMapper: {
     "^(\\.{1,2}/.*)\\.js$": "$1",
     "^TiFShared/(.*)\\.js$": "<rootDir>/node_modules/TiFShared/$1"
-  },
+  }
 }
