+++ GeocodingLambda/tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "extends": "./node_modules/TiFShared/tsconfig.json",
   "compilerOptions": {
     "module": "esnext",
     "target": "es2020",
